+++ lib/domain/repositories/pharmacy_repository.dart
@@ -41,7 +41,7 @@ abstract class PharmacyRepository {
   Future<Either<Failure, ProductDTO>> updatePharmacyProductById({
     required int productId,
     String? status,
-    int? scanCount,
+    double? scanCount,
     int? defective,
     int? surplus,
     int? underachievement,
